refactor: move revenue service provider to src/Revenue
